Update README and make some fixes for pagination and other bugs
