+++ buildAndPush.sh
@@ -6,7 +6,7 @@ echo
 echo "Using docker hub repository ${repo}"
 echo
 echo "+============================================"
-echo "| START: ATLAS NEWS SEARCH MULTI-ARCH BUILD"
+echo "| START: ATLAS NEWS DEMO MULTI-ARCH BUILD"
 echo "+============================================"
 echo
 
@@ -16,6 +16,6 @@ abbrvhash=${datehash: -8}
 echo 
 echo "Building container using tag ${abbrvhash} and buildx"
 echo
-docker buildx create --name newssearchbuilder --use
+docker buildx create --name builder --use
 docker buildx build --platform linux/amd64,linux/arm64 -t ${repo}:latest -t ${repo}:${abbrvhash} --push .
-docker buildx rm newssearchbuilder
+docker buildx rm builder
