+++ example.env
@@ -10,3 +10,4 @@ FIREWORKS_API_KEY="<put key here>"
 NOMIC_API_KEY="<put key here>"
 PROVIDER="openai | azure_openai | vectorservice | mistral | fireworks | nomic"
 EMBEDDING_DIMENSIONS=<integer>
+EMBEDDING_MODEL="<name of model. E.g. text-ada-002>"
