+++ runLocal.sh
@@ -1,8 +1,3 @@
-# source .env
-# cd frontend
-# npm run build
-# npm start
-
 source .env
 . venv/bin/activate && \
 supervisord -c supervisord.conf
